Move calculate_tanimoto_scores_unique_inchikeys to CalculateScoresBetweenAllIonmodes.py
