+++ ms2deepscore/benchmarking/calculate_scores_for_validation.py
@@ -1,18 +1,10 @@
 import os
 from typing import List
-import numpy as np
-import pandas as pd
 import torch
-from matchms.filtering.metadata_processing.add_fingerprint import \
-    _derive_fingerprint_from_smiles
-from matchms.similarity.vector_similarity_functions import \
-    jaccard_similarity_matrix
 from matchms.Spectrum import Spectrum
-from tqdm import tqdm
 from ms2deepscore import MS2DeepScore
+from ms2deepscore.benchmarking.CalculateScoresBetweenAllIonmodes import calculate_tanimoto_scores_unique_inchikey
 from ms2deepscore.models import load_model
-from ms2deepscore.train_new_model.inchikey_pair_selection import \
-    select_inchi_for_unique_inchikeys
 from ms2deepscore.utils import save_pickled_file
 
 
@@ -61,41 +53,6 @@ def calculate_true_values_and_predictions_for_validation_spectra(positive_valida
     del ms2deepscore_model
 
     return true_values_collection, predictions_collection
-
-
-def calculate_tanimoto_scores_unique_inchikey(list_of_spectra_1: List[Spectrum],
-                                              list_of_spectra_2: List[Spectrum],
-                                              fingerprint_type="daylight",
-                                              nbits=2048
-                                              ):
-    """Returns a dataframe with the tanimoto scores between each unique inchikey in list of spectra"""
-
-    def get_fingerprint(smiles: str):
-        fingerprint = _derive_fingerprint_from_smiles(smiles,
-                                                      fingerprint_type=fingerprint_type,
-                                                      nbits=nbits)
-        assert isinstance(fingerprint, np.ndarray), f"Fingerprint could not be set smiles is {smiles}"
-        return fingerprint
-
-    if len(list_of_spectra_1) == 0 or len(list_of_spectra_2) == 0:
-        raise ValueError("The nr of spectra to calculate tanimoto scores should be larger than 0")
-
-    spectra_with_most_frequent_inchi_per_inchikey_1, unique_inchikeys_1 = \
-        select_inchi_for_unique_inchikeys(list_of_spectra_1)
-    spectra_with_most_frequent_inchi_per_inchikey_2, unique_inchikeys_2 = \
-        select_inchi_for_unique_inchikeys(list_of_spectra_2)
-
-    list_of_smiles_1 = [spectrum.get("smiles") for spectrum in spectra_with_most_frequent_inchi_per_inchikey_1]
-    list_of_smiles_2 = [spectrum.get("smiles") for spectrum in spectra_with_most_frequent_inchi_per_inchikey_2]
-
-    fingerprints_1 = np.array([get_fingerprint(spectrum) for spectrum in tqdm(list_of_smiles_1,
-                                                                              desc="Calculating fingerprints")])
-    fingerprints_2 = np.array([get_fingerprint(spectrum) for spectrum in tqdm(list_of_smiles_2,
-                                                                              desc="Calculating fingerprints")])
-    print("Calculating tanimoto scores")
-    tanimoto_scores = jaccard_similarity_matrix(fingerprints_1, fingerprints_2)
-    tanimoto_df = pd.DataFrame(tanimoto_scores, index=unique_inchikeys_1, columns=unique_inchikeys_2)
-    return tanimoto_df
 
 
 def get_tanimoto_score_between_spectra(spectra_1: List[Spectrum],
